Make pathfinding more consistent
Pathfinding no longer takes into account units in the room, unless a
unit is detected as being stuck, in which case it is allowed to know
where the other units are in order to get itself unstuck.

This makes pathfinding much more consistent now that we are caching cost
matrices, because there is no guarantee that the location of a
non-static entity like a unit is correct in the cost matrix any more.